Move changed integration tests into mapped suites
